requirements.txt
azure-functions
PyMuPDF
requests
psycopg2-binary
